src/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachelib==0.10.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
dataclasses-json==0.5.14
Deprecated==1.2.14
Django==3.2.2
django-cors-headers==4.0.0
django-webpack-loader==2.0.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.1.2
frozenlist==1.4.0
fsspec==2023.10.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httptools==0.6.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
josepy==1.13.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.330
langsmith==0.0.57
llama-index==0.8.59
MarkupSafe==2.1.2
marshmallow==3.20.1
mozilla-django-oidc==1.2.4
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.25.1
openai==0.27.8
openpyxl==3.0.9
outcome==1.2.0
packaging==23.2
pandas==2.0.3
pandas-stubs==2.0.2.230605
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyOpenSSL==23.2.0
pypdf==3.17.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.23
sqlparse==0.4.4
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
types-pytz==2023.3.0.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.6
uvicorn==0.24.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==2.3.1
whitenoise==5.3.0
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0
